Move each test next to it's module
